% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cda3.R
\name{logit_cda_r}
\alias{logit_cda_r}
\title{Optimize a logistic regression model by coordinate descent algorithm using a design matrix with R}
\usage{
logit_cda_r(X_tilde, y, lambda, R, init.beta, delta, maxit, eps, warm, strong,
  sparse)
}
\arguments{
\item{X_tilde}{standardized matrix of explanatory variables}

\item{y}{vector of objective variable}

\item{lambda}{lambda sequence}

\item{R}{matrix using exclusive penalty term}

\item{init.beta}{initial values of beta}

\item{delta}{ratio of regularization between l1 and exclusive penalty terms}

\item{maxit}{max iteration}

\item{eps}{convergence threshold for optimization}

\item{warm}{warm start direction: "lambda" (default) or "delta"}

\item{strong}{whether use strong screening or not}

\item{sparse}{whether use sparse matrix or not}
}
\value{
standardized beta
}
\description{
Optimize a logistic regression model by coordinate descent algorithm using a design matrix with R
}
\keyword{internal}
