% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cda3.R
\name{soft_threshold}
\alias{soft_threshold}
\title{soft thresholding function}
\usage{
soft_threshold(z, g)
}
\arguments{
\item{z}{z}

\item{g}{gamma}
}
\value{
value
}
\description{
soft thresholding function
}
\keyword{internal}
