% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{softThresholdC}
\alias{softThresholdC}
\title{soft thresholding function}
\usage{
softThresholdC(z, g)
}
\arguments{
\item{z}{z}

\item{g}{gamma}
}
\value{
value
}
\description{
soft thresholding function
}
\keyword{internal}
